Rename the `module` action `modulecmd`
It is actually a wrapper to `modulecmd` and should be named as such.
Instead, `flight` defines a bash `module` method that hooks into the
`modulecmd` action.

The `module` action file is still remains as a placeholder to the CLI
only. It should not be ran as this is up to flight direct