Package up FlightSyncer with ActiveModel
The FlightSyncer library will be built using ActiveModel. ActiveModel is
packaged into the project using `bundle package` into the `vendor` dir.

The vendor directory is .gitignored to prevent the repo getting to big.
However the `.bundle/config` files are now checked in as they have a
large impact on how bundler operates.

Also the blank FlightSyncer files have been created, one in lib and the
other in libexec