Define a method to convert the CacheFile to a hash
The `to_h` will convert the `CacheFile` to a hash. This will make saving
the data to `yaml` easier.

To ensure the correct keys are picked up, they are defined in a
constant. The content of the file is not stored in the hash as it will
be pulled from a URL (TBD)