Require the file content to save to the cache
The file URL is not going to be supplied with the metadata and thus
prevents the manifest being tied to server configuration.

Instead, the URL can be inferred from the `identifier` and env vars.
This does require the server to store the files correctly.

Thus the file content is required to preform the `save_to_cache` action.
The `SyncManifest` class temporarily caches the content in a hash and
saves them when `save` is called.