Define a method to save from the cache
Once the metafile is extracted from the manifest, it can be used to
save the file in place. At this point in time, it ignores all the file
permissions and owner/group information.

All it does is download the file content and save it to the location as
specified by the metadata