Allow a custom path to be specified from the CLI
When a file is added to the cache, the orginal file might not be in its
final sync location. As such a `--path` flag has been added to the CLI

The MetaFile had to be updated so parameters can be assigned when the
object is initialized. Then parameters are only inferred from the
original file IFF it hasn't been set on initialize