Implement the add commands to manage the client
The `add` command no longer interacts with the server at all. All server
commands are listed under the cache `subcommand`.

Instead the `add` commands will store a list of files/ groups on the
client. This will tell the client which files it needs to to sync

A new `YAML_File` struct has been created to abstract away the file
handling. The `update` method will by default return a Hashie::Mash
which is saved at the end.