Sync all the groups and files at the same time
Now that the list of files/ groups to be synced are stored on the
client, it is easy enough to sync them all at once.

However, the groups might contain duplicate identifiers which need to be
filtered out. As such it is easier to grap the file/identifiers within
the groups, filter the duplicates and then generate the metafiles.

As such the `metafiles_in_group` method is obsolete and has been removed