Add a command to preform the package snapshot
The `anvil snapshot` command will now run the snapshot. It `exec` into
the command so that the user continues to get the output to stdout and
stderr.

This replaces the question on which address to build the package on

NOTE: The `rake` command is still executed due to issues with requiring
the rails app