Manually preform the snapshot steps
A lot of the snapshot code will be moved into this file (eventually). To
facilitate this, the snapshot has been broken down into steps. Some of
these will remain rake tasks, others may change

NOTE: Some of the snapshot code is only defined within the
`packages:snapshot` task. As this task is going to be deprecated, this
code is currently being skipped.