Add a new ANVIL environment variable
Due to the large amount of code that has been ported from anvil, the
array of environment variables are no longer required.

The only remaining task is `packages:import` (ignoring `db:drop`)
This task will remain in ANVIL as it needs to connect to rails

However it only requires the path to the directory the packages are
stored in. This path will be set by `ANVIL_IMPORT_DIR`. The other env
vars will eventually be removed