Create a `postgres` forge package
This package is broadly based on the `flight-cache` package. It uses the
original build code used to create the original `extract-postgres.sh`
file.

The major change is that postgres will be installed within the
`/opt/flight-direct` directory. This keeps it separate from the system
version of postgres (if it is installed). In theory, this db could run
along side the system version (Maybe?)

The service has been renamed `postgres-flight` so it does conflict with
the system version. The `postgres` user is still the owner of the db