Use the postgres package instead of the S3 tarball
Now there is a specific postgres package for flight direct, the old
tarball is now obsolete. Instead, the rails app should connect to the
internal version.

This does require compiling the `pg` gem with this particular version.
As this can only be done at install time, a two step process is
required.

At build time, the gems are cached ('bundle package') and `pg` is
configured to use the flight-direct posgres.

Then at install time, the cache is used to compile all the gems. As
postgres is a package dependency, it has already been installed. This
allows `pg` to be built correctly